Reduce mem used in RunningAverage class

See merge request setsm/setsm!43

+++ Typedefine.hpp
@@ -1291,7 +1291,7 @@ class RunningAverage {
 
     private:
         float v;
-        long n;
+        int n;
 
 };
